requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
python-dotenv
openai
requests
pydantic

